ver0.1/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0 